using promoted type as argument for va_arg fixes crashes with gcc-3.2
Sun Feb  9 17:56:07 2003  Stefan Westerfeld  <stefan@space.twc.de>

	* sfiglue.c: using promoted type as argument for va_arg fixes
	crashes with gcc-3.2

svn path=/trunk/; revision=873

+++ sfi/sfiglue.c
@@ -326,7 +326,7 @@ sfi_glue_call_valist (const gchar *proc_name,
 	  sfi_seq_unref (seq);
 	  return NULL;
 	}
-      arg_type = va_arg (var_args, guint8);
+      arg_type = va_arg (var_args, guint);
     }
   if (seq)
     {
